Cleanup: cast bytesProcessed to quint64 instead of truncating bytesToWrite
Followup to 9d5aaec: code-review flagged static_cast<int>(bytesToWrite)
as a latent overflow risk. For cut segments >2 GB, casting the quint64
down to int would produce a negative value, making the EOS guard fire
spuriously. Cast the small bounded bytesProcessed up to quint64 instead
— canonical wider-domain comparison, no truncation risk.

+++ avstream/ttmpeg2videostream.cpp
@@ -596,7 +596,7 @@ void TTMpeg2VideoStream::transferCutObjects(TTVideoHeader* startObject, TTVideoH
         // already (no more chunks to read), don't seekBackward — that creates
         // an infinite loop where each iteration re-reads the same trailing
         // bytes without making progress. Just process and exit the inner loop.
-        if (bytesProcessed >= static_cast<int>(bytesToWrite)) {
+        if (static_cast<quint64>(bytesProcessed) >= bytesToWrite) {
           break;
         }
         stream_buffer->seekBackward(watermark);
